+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed bug in `compas.geometry.is_coplanar`.
 * Fixed bug in `compas.datastructures.mesh_merg_faces`.
 * Fixed bug in `compas.robots.RobotModel.add_link`.
+* Fixed bug in `compas.datastructures.Volmesh.cell_to_mesh`.
 
 ### Removed