+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `remap_values` to `compas_utilities`.
+
 ### Changed
 
 * Fixed bug in `compas.datastructures.Network.delete_node`.